Make a lot of the numbers be 8 bits to reduce memory usage

+++ src/solver.rs
@@ -8,11 +8,11 @@ impl Solver {
         let mut lines = input.lines();
 
         let line = lines.next().expect("Couldn't read from input");
-        let numbers_count: u32 = line.trim().parse().expect("Couldn't parse input as number");
-        let grid_side = (numbers_count as f32).sqrt().ceil() as u32;
+        let numbers_count: u8 = line.trim().parse().expect("Couldn't parse input as number");
+        let grid_side = (numbers_count as f32).sqrt().ceil() as u8;
 
         let line = lines.next().expect("Couldn't read from input");
-        let zero_number: i32 = line.trim().parse().expect("Couldn't parse input as number");
+        let zero_number: i8 = line.trim().parse().expect("Couldn't parse input as number");
 
         let mut numbers_buf = String::new();
 
@@ -21,7 +21,7 @@ impl Solver {
             numbers_buf.push('\n');
         }
 
-        let split_numbers: Vec<u32> = numbers_buf
+        let split_numbers: Vec<u8> = numbers_buf
             .split_whitespace()
             .map(|x| x.parse().expect("Couldn't parse input as number"))
             .collect();
